--- a/doctest/Main.hs
+++ b/doctest/Main.hs
@@ -1,8 +1,0 @@
-import Test.DocTest
-
-main :: IO ()
-main = doctest [
-    "-isrc"
-  , "-XOverloadedStrings"
-  , "src/URI/ByteString.hs"
-  ]