--- a/uri-bytestring.cabal
+++ b/uri-bytestring.cabal
@@ -35,7 +35,6 @@ library
       attoparsec       >= 0.10    && < 0.13
     , base             >= 4.6     && < 4.8
     , bytestring       >= 0.10.2.0 && < 0.11
-    , profunctors      >= 4.0 && < 5.0
 
   hs-source-dirs:      src
   default-language:    Haskell2010