… bench/Main.hs
… test/URI/ByteString/Arbitrary.hs
… test/URI/ByteString/LensTests.hs
… test/URI/ByteStringTests.hs