Since requests (at least in snap) contain only path, query, and fragment, I would like to have parser and serializer functions that give me that.

1. If you are generally ok with the idea, I will can test cases.
2. Not sure about the name.  I think `relative` is for rootless paths.  Is there a better technical term for schema-and-authority-free urls?

thanks,
m.