@mf59816 2015-04-22
Oh, and note that I have based this branch on `lens`.  Merge with care!