@MichaelXavier 2015-04-22
I looked at RFC3986 and indeed relative URIs are a thing, but they are called "relative-ref" now. Just search this document for "relative-ref": https://www.ietf.org/rfc/rfc3986.txt

It feels awkward to me to have to supply information you potentially don't have (authority/scheme) to use a relative parser when the spec says it can stand on its own, so here's what I recommend:

1. Review the lens branch and if everything looks good, I'll
   merge. Merge that into your branch so you have an up-to-date field
   naming scheme.
2. Add a `RelativeRef` type to the `URI.ByteString.Types` module and
   export it from `URI.ByteString`.
3. Have your parser parse that type and remove any extraneous
   arguments.
4. Have `uriParser` be implemented in terms of this `relativeRef`
   parser like you do now and just have it pull apart the
   `RelativeRef` and build a `URI` with it.
5. Update the serialize functions in this same way.
6. Add a few tests and a benchmark for the new code.