@MichaelXavier 2015-04-22
Oh and just a heads up in case you got started, I was a bit hasty with the lens branch and pushed the change before discussing it with my team. I've got sign off with what you'll see on lens right now: the lens module isn't exported, all type fields don't have the _ in them. All lenses are postfixed with L.