@mf59816 2015-04-22
Yes, I was briefly confused by that, but now I'm rebased on 74fc43f1.  Thanks!  I like that better, too.