@mf59816 2015-04-22
Rebased to tip of lens branch.  I'll post the criterion figures in a moment.