@mf59816 2015-04-22
```
before (lens branch)
====================

benchmarking parsing/Network.URI.parseURI
time                 6.649 μs   (6.638 μs .. 6.661 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 6.680 μs   (6.655 μs .. 6.740 μs)
std dev              121.2 ns   (58.54 ns .. 220.3 ns)
variance introduced by outliers: 17% (moderately inflated)

benchmarking parsing/URI.ByteString.parseURI strict
time                 2.244 μs   (2.235 μs .. 2.255 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.234 μs   (2.229 μs .. 2.248 μs)
std dev              26.89 ns   (14.04 ns .. 49.38 ns)

benchmarking parsing/URI.ByteString.parseURI lax
time                 2.355 μs   (2.349 μs .. 2.363 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.371 μs   (2.362 μs .. 2.383 μs)
std dev              34.15 ns   (24.31 ns .. 44.56 ns)
variance introduced by outliers: 13% (moderately inflated)

benchmarking serializing/URI.ByteString.serializeURI
time                 1.838 μs   (1.836 μs .. 1.841 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.844 μs   (1.839 μs .. 1.852 μs)
std dev              20.23 ns   (13.17 ns .. 30.57 ns)

Benchmark bench: FINISH


after (this branch)
===================

benchmarking parsing/Network.URI.parseURI
time                 6.727 μs   (6.693 μs .. 6.759 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 6.708 μs   (6.676 μs .. 6.744 μs)
std dev              116.2 ns   (103.7 ns .. 135.5 ns)
variance introduced by outliers: 16% (moderately inflated)

benchmarking parsing/URI.ByteString.parseURI strict
time                 2.314 μs   (2.303 μs .. 2.328 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 2.371 μs   (2.342 μs .. 2.422 μs)
std dev              129.5 ns   (94.12 ns .. 172.2 ns)
variance introduced by outliers: 68% (severely inflated)

benchmarking parsing/URI.ByteString.parseURI lax
time                 2.440 μs   (2.400 μs .. 2.510 μs)
                     0.996 R²   (0.991 R² .. 1.000 R²)
mean                 2.442 μs   (2.422 μs .. 2.489 μs)
std dev              100.7 ns   (36.81 ns .. 201.6 ns)
variance introduced by outliers: 55% (severely inflated)

benchmarking parsing/URI.ByteString.parseRelativeRef strict
time                 1.540 μs   (1.513 μs .. 1.580 μs)
                     0.998 R²   (0.997 R² .. 1.000 R²)
mean                 1.523 μs   (1.516 μs .. 1.539 μs)
std dev              34.43 ns   (18.41 ns .. 66.04 ns)
variance introduced by outliers: 27% (moderately inflated)

benchmarking parsing/URI.ByteString.parseRelativeRef lax
time                 1.694 μs   (1.672 μs .. 1.729 μs)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 1.730 μs   (1.705 μs .. 1.763 μs)
std dev              90.62 ns   (78.00 ns .. 106.1 ns)
variance introduced by outliers: 67% (severely inflated)

benchmarking serializing/URI.ByteString.serializeURI
time                 1.960 μs   (1.892 μs .. 2.059 μs)
                     0.986 R²   (0.977 R² .. 0.996 R²)
mean                 1.949 μs   (1.899 μs .. 2.041 μs)
std dev              221.4 ns   (145.7 ns .. 334.0 ns)
variance introduced by outliers: 91% (severely inflated)

benchmarking serializing/URI.ByteString.serializeRelativeRef
time                 1.704 μs   (1.692 μs .. 1.714 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 1.715 μs   (1.698 μs .. 1.744 μs)
std dev              69.19 ns   (32.61 ns .. 120.1 ns)
variance introduced by outliers: 54% (severely inflated)

Benchmark bench: FINISH
```