@mf59816 2015-04-22
If you want to be really picky about it, you may be worried about section 4.2 in the rfc:

```
   A path segment that contains a colon character (e.g., "this:that")
   cannot be used as the first segment of a relative-path reference, as
   it would be mistaken for a scheme name.  Such a segment must be
   preceded by a dot-segment (e.g., "./this:that") to make a relative-
   path reference.
```

In contrast to that:

```haskell
x = toLazyByteString . serializeRelativeRef $ RelativeRef
    Nothing "./this:that/wef/eofij" (Query [("1", "2")]) Nothing
y = parseRelativeRef laxURIParserOptions $ cs x
```

`y` evaluates to `Left (OtherError "Failed reading: mzero")`.  Without the `./`, parsing fails as specified in the RFC.  (`cs` is from string-conversions.)

Please let me know if you want me to look into it.  I personally could happily live with this.