@MichaelXavier 2015-04-22
I think we should try to take the high road here if possible. Since this library makes a point of being an RFC-compliant parser, we should stick to it. If you wouldn't mind trying to get this to conform to the spec that would be great. If it looks to hairy let me know and I'll give it a try.