@mf59816 2015-04-23
I knew I shouldn't have brought it up.  ;-)

Then again, fixing this obscure corner case made me discover that rootless relative refs were still broken.  That I would have minded soon.