@mf59816 2015-04-24
I think this is ready.

(Would you consider including this in release 0.1?  When can I hope to see that on hackage?  You've been very helpful and responsive, and I hope I'm not being impatient now.  I'd just like to make plans.  Thanks!)