@MichaelXavier 2015-04-24
This looks great! Thanks for digging into the edge cases. I'll cut a release and add you to the contributors. Really appreciate it!