@MichaelXavier 2015-04-24
Welp it looks like hackage is rejecting package with -Werror flags so I'm going to try to get that sorted out with them. Really annoying. I recommend installing this locally if you need something right away. If I can't get any progress on it, i'll just take off the flags for the release I guess.