@MichaelXavier 2015-04-27
Rather than trying to fight the law and fix cabal and hackage before uploading, I nixed the error flag, built and uploaded it. 0.1 is up.