@mf59816 2015-04-27
great.  thanks!  that was fun  (-:

(Found your fight with the law; for reference: https://github.com/haskell/hackage-server/issues/259.)